+++ tests/ByteRangeSet.cpp
@@ -17,6 +17,7 @@
 
 #include "../src/platform.hpp"
 #include <gtest/gtest.h>
+#include <stdint.h>
 #include <stdio.h>
 
 #include "../src/ByteRangeSet.hpp"
@@ -33,7 +34,7 @@ using namespace REHex;
 std::ostream& operator<<(std::ostream& os, const ByteRangeSet::Range& range)
 {
 	char buf[128];
-	snprintf(buf, sizeof(buf), "{ offset = %zd, length = %zd }", range.offset, range.length);
+	snprintf(buf, sizeof(buf), "{ offset = %jd, length = %jd }", (intmax_t)(range.offset), (intmax_t)(range.length));
 	
 	return os << buf;
 }
